Moved Layer and Color to Model package
