+++ src/main/java/C7/IO/LayerIO.java
@@ -1,7 +1,7 @@
 package C7.IO;
 
-import C7.Color;
-import C7.Layer.Layer;
+import C7.Model.Color;
+import C7.Model.Layer.Layer;
 import javafx.scene.image.Image;
 import javafx.scene.image.PixelReader;
