+++ src/main/java/C7/Project.java
@@ -1,6 +1,7 @@
 package C7;
 
-import C7.Layer.ILayer;
+import C7.Model.Color;
+import C7.Model.Layer.ILayer;
 
 import java.util.ArrayList;
