+++ src/test/java/C7/Model/Tools/CalligraphyTest.java
@@ -1,6 +1,6 @@
 package C7.Model.Tools;
 
-import C7.Color;
+import C7.Model.Color;
 import C7.Model.Vector.Vector2D;
 import org.junit.Assert;
 import org.junit.Test;
